% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammabustr.R
\name{gammabustr}
\alias{gammabustr}
\title{Gamma Prior Elicitation}
\usage{
gammabustr(
  mode,
  pct,
  bound,
  dir = "lt",
  ndraw = 10000,
  beta_upper = 100,
  plot = FALSE
)
}
\arguments{
\item{mode}{The most likely value}

\item{pct}{Percent confidence in boundary statement}

\item{bound}{Boundary value}

\item{dir}{Direction from boundary value. Valid values are "lt" and "gt", indicating "less than" and "greater than".}

\item{ndraw}{Number of random draws from distribution}

\item{beta_upper}{Max beta value to search}

\item{plot}{Logical whether to plot final distribution}
}
\value{
A named numeric vector with alpha and beta values for a gamma distribution. If plot == TRUE, also plots a gamma distribution with those parameters and indicates the (analytically calculated) mode by a vertical dotted line. Discrepancy between the location of the dotted line and the peak of the curve is an indication of poor convergence.
}
\description{
Returns alpha and beta values for a gamma distribution given statements of the most likley value and a confidence statement for a boundary value, either less than or greater than.
}
\details{
This function mathematically formalizes expert knowledge about a gamma distributed variable. Experts provide a most likely value, which is hard coded as the mode of the distribution, as well as a confidence statement about whether the value is greater than or less than some boundary values. For example, an expert might say "I believe the average height of women in Albuquerque, New Mexico is 1.6 meters, and I'm 95 percent sure it's less than 1.8 meters." The function then performs a parameter search via Brent's method to determine the shape (alpha) and scale (beta) parameters of a gamma distribution with the location and scale specified by the expert.
}
\examples{
gammabustr(mode = 0.5, pct = 0.95, bound = 2, dir = "lt")

}
